+++ native/scripts/build_pod.sh
@@ -13,6 +13,7 @@ SRCS=(
   "${ROOT}/src/bus.c"
   "${ROOT}/src/frame.c"
   "${ROOT}/src/pipeline.c"
+  "${ROOT}/src/thumbnail.c"
   "${ROOT}/src/xhvp_player.c"
   "${ROOT}/src/xhvp_ffi_keep.c"
 )